core_app/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
SQLAlchemy==2.0.30
alembic==1.13.2
pydantic[email]==2.7.1
pydantic-settings==2.3.4
bcrypt==4.1.3
PyJWT==2.8.0
python-dotenv==1.0.1
httpx==0.27.0
pytest          # async MongoDB driver (built on PyMongo)
pymongo==4.8.0         # pinned by motor; explicit for tools/index helpers
motor>=3.3,<4.0
beanie>=1.25,<2.0
aiosmtplib==2.0.2
zxcvbn-python==4.4.24
PyJWT==2.8.0
httpx==0.27.0
